requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
first==2.0.2
gunicorn==20.0.4
idna==2.10
numpy==1.19.1
opencv-python==4.4.0.42
pip-tools==1.8.0
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.0
urllib3==1.25.10
whitenoise==5.2.0
